Suppress memory tools warning if tests will be skipped
I'd prefer to inspect the test itself to see if it will be skipped
individually, but this warning is implemented in a common location so
that it is printed once, instead of once for each defined test. Looking
at the source variable, `AMENT_RUN_PERFORMANCE_TESTS`, should give us
the same behavior.

Signed-off-by: Scott K Logan <logans@cottsay.net>