Fix a warning when building on Ubuntu Noble. (#26)
In particular, gcc 13.2 was complaining that we were
accessing a pointer after a free.  And that was technically
true; the calls to DoNotOptimize(ptr) were after the
free.  Just move this before the free (when the ptr is still
valid) to remove the warning, but still ensure that we don't
optimize the pointer away.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>
(cherry picked from commit e406c3cf4a352ab3f89c43fcebe7503781be5905)